+++ example/lib/screens/forms/auto_suggest_box.dart
@@ -236,7 +236,10 @@ if (isOverlayVisible) {
                     onSelected: (item) {
                       setState(() => selectedCat = item.value);
                     },
-                    onOverlayVisibilityChanged: (_) => setState(() {}),
+                    onOverlayVisibilityChanged: (visible) {
+                      debugPrint('Overlay is visible: $visible');
+                      setState(() {});
+                    },
                   ),
                 ),
                 Flexible(